requirements.txt
os 
mysql-connector
bcrypt